requirements.txt
nest_asyncio
aiohttp
python-telegram-bot==20.0
